Separating plotting from writing to file
